LoopVariables has separate source files for constructor and destructor
